Save WP page associations properly during installation
The 'name' attribute of the <select> elements on the Pages tab of the wizard
were previously hardcoded with the incorrect value (they were all called
"...-blogs-..."). As a result, it was impossible to associate components
with existing pages during installation (except for the Blogs component). This
changeset replaces the names with the correct values.

Fixes #4326

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6169 cdf35c40-ae34-48e0-9cc9-0c9da1808c22