Declare $active_components property in BuddyPress class
The addition of the __set() and __get() magic methods on BuddyPress() mean that
any property not explicitly declared in the class definition will be set
dynamically as an overloaded property. In the case of $active_components, this
overloading causes a problem with pre-BP-1.5-era plugins that registered
themselves directly in the $bp->active_components array, because overloaded
properties cannot be modified in this way.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6334 cdf35c40-ae34-48e0-9cc9-0c9da1808c22