Fix incorrect function call in bp_total_mention_count_for_user()
It appears that, since bp_total_mention_count_for_user() was introduced in
r2287, it has echoed the output of the incorrect function (_favorite_count_
rather than _mention_count_). It was probably never noticed because BP never
uses the echo function natively; we always use the _get_ function in the
context of a sprintf().

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7057 cdf35c40-ae34-48e0-9cc9-0c9da1808c22