Use is_callable() to validate screen_function callbacks before hooking in bp-core-buddybar.php
We were previously using is_object() and other custom verification techniques
to hook screen functions in bp_core_new_nav_item() and elsewhere. These
techniques are limited in that they allow certain sorts of uncallable strings
to pass through. We also should not be reinventing the wheel.

Fixes #4794

Props r-a-y

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7171 cdf35c40-ae34-48e0-9cc9-0c9da1808c22