Show member Profile subnav for all users, not just when viewing one's own profile
The logic for hiding others' Profile subnavs comes from the fact that, as a
rule, the only subnav item available when viewing another person's profile is
View/Public. But this causes problems when plugins or themes try to add
additional tabs to user Profiles.

By removing the bp_is_my_profile() check hardcoded in the template, we rely on
the bp_core_new_subnav_item() logic in setup_nav() to do the necessary access
control (which it was already doing).

Fixes #5039

Props r-a-y

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7211 cdf35c40-ae34-48e0-9cc9-0c9da1808c22