Check username column first before sending a PM to a recipient.
Previously, when sending a private message, a check was made against the
user ID column first to determine who the recipient is.

The problem with that is if:
  a. A site uses numeric user_login / user_nicenames only; and if
  b. A site has a large userbase

That the recipient might be calculated incorrectly because of the check
against the ID column first leading to unintended private messages being
sent to the wrong user(s).

This commit fixes this issue by reversing the look up so the user_login /
user_nicename DB column is checked first, followed by the ID column being
checked as a fallback if there is no initial match.

Fixes #5151.

Props DennisSmolek.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7462 cdf35c40-ae34-48e0-9cc9-0c9da1808c22