XProfile: Show the placeholder text for all select box profile fields.
Prior to BP 1.5, the "----" placeholder item for select boxes was shown for both optional and mandatory fields. In BP 1.5 in r4831, the behaviour was changed to only show the placeholder item for optional profile fields. With retrospect, that change didn’t make much sense.

This commits restores the original behaviour, showing the placeholder text for all select box profile fields. A user should be forced to pick an option for a required profile field, and not allow the web browser to automatically select the first option in the <select> element.

See #3378, #3698, and #5184. Props haykayltduk, hnla, DJPaul at WordCamp London contributor day!

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7599 cdf35c40-ae34-48e0-9cc9-0c9da1808c22