Install default options at installation time and at upgrade to 2.0
When BuddyPress is installed, all basic config options should be saved to the
database, even when the setting is empty (0). This ensures that the options
will be pre-fetched by the first call to wp_load_alloptions() and stored in the
non-persistent cache, removing the necessity to make additional failed queries
to the wp_options table.

This changeset also includes logic to add missing default options to existing
default installations when upgrading to 2.0.

Fixes #5385

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7904 cdf35c40-ae34-48e0-9cc9-0c9da1808c22