Pre-fetch blogmeta when in blog loops
This reduces overhead when querying for blog meta in the context of a
bp_has_blogs() template loop. Mirrors similar functionality in the activity,
groups, and xprofile components.

Also introduces an update_meta_cache parameter to the bp_has_blogs() stack, so
that plugin/theme authors can disable this pre-fetching.

Fixes #5398

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7918 cdf35c40-ae34-48e0-9cc9-0c9da1808c22