Overhaul the way that individual group objects are cached
The existing implementation was broken in a number of ways. Most critically,
user-specific data (like is_member) was being stored in the persistent object
cache without respect to user.

This refactor excludes the user-specific data from being cached along with the
group object. It also reworks the way that items are keyed in the cache, for
greater consistency with WordPress and the other BP components.

See #5407

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7956 cdf35c40-ae34-48e0-9cc9-0c9da1808c22