Do not query for a group unnecessarily if we're on a group page.
If we're on a group page, the group object is already queried in
the BP_Groups_Component::setup_globals() method.  However,
BP_Groups_Template requeries for the group object even when we're on a
single group page.

This commit changes BP_Groups_Template to use the already-queried
group object to prevent these unnecessary queries.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8050 cdf35c40-ae34-48e0-9cc9-0c9da1808c22