Auto-fetch URL search parameters in bp_has_groups() before parsing $args array
Our previous method, of getting the search term out of $_REQUEST, had the
effect of overriding any search terms that were passed to the bp_has_groups()
function. Among other problems, this made it impossible to get an unfiltered
group list when looking at a page that was using an 's' parameter for something
else.

Fixes #5456

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8116 cdf35c40-ae34-48e0-9cc9-0c9da1808c22