Introduce Sign Ups Management
In Users Administration Screen, a new view is now available to manage the pending accounts of a site or of the network of sites. The following actions are supported:
- Resend the activation email
- Delete the pending account
- Activate the pending account
The corresponding bulk actions are also supported. A search box is available in order to let the administrator easily find some specific pending accounts.

The registration process have also been modified so that multisite and regular configs handles it in a similar way. A mechnanism is in place to ensure plugin backward compatibility concerning the regular configs.

See #5374

props boonebgorges, imath

Fixes #4651

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8119 cdf35c40-ae34-48e0-9cc9-0c9da1808c22