Groups: Add backward compatibility when fetching list of admins/mods.
In r8056, the change was made so a group does not fetch its extra data
like grabbing the list of group admins and mods by default.

However, if a group template loop was created with the 'populate_extras'
parameter set to false and either the bp_group_list_admins() and
bp_group_list_mods() functions were used in the loop, these functions
would not output the correct list.

To solve this, this commit manually fetches the group admins / mods if
the 'populate_extras' flag in the group template loop is set to false.

Fixes #5411

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8168 cdf35c40-ae34-48e0-9cc9-0c9da1808c22