Don't attempt to overload add_root property in bp_core_add_root_component()
The add_root array is defined only for backward compatibility purposes in the
BuddyPress global object, and thus is overloaded rather than defined in the
constructor. This means that adding items directly to the array is forbidden
on recent versions of PHP. As a workaround, we add the item as a separate array
and then swap the array, which is allowed and will not result in notices.

Fixes #5588

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8323 cdf35c40-ae34-48e0-9cc9-0c9da1808c22