Check bp_disable_account_deletion() setting in screen function rather than bp_core_delete_account() business function
The purpose of the disable-account-deletion setting is to prevent normal (ie
non-admin) BP users from deleting their own accounts through the front-end
interface. Locating the settings check in the business function
bp_core_delete_account() therefore caused problems when deleting accounts in
areas that were not meant to be covered by the setting, such as when accounts
are deleted in the Dashboard by an admin, or when the BP_Signup::delete()
method is called. This caused a particular problem in BP 2.0, because the
deletion of signups was no longer correctly cleaning up other data from the
database (xprofile, usermeta, etc).

Fixes #5575

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8324 cdf35c40-ae34-48e0-9cc9-0c9da1808c22