Check groups loop data when using `bp_group_is_user_banned()`.
Previously, `bp_group_is_user_banned() did not reference the data already
queried in the groups loop via `BP_Groups_Group::get_group_extras()`.  As a
result, if a user without the `'bp_moderate'` cap was logged in and viewing
the group directory, this added twenty additional DB queries to the page.

Commit also standardizes the return value of `bp_group_is_user_banned()` to
a boolean.  Previously, it was possible for the function to return either a
string of the integer, null or a boolean.

Fixes #5705.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8531 cdf35c40-ae34-48e0-9cc9-0c9da1808c22