xProfile: fix profile fields not rendering if their value is "0".
This was caused by a reliance on the `empty()` function in a few places, which was causing us to
not handle "0" (both as int and string) as a valid field value. In turn, this was causing the
profile field template loop to skip rendering the affected field.

Fixes #5731, props DJPaul and r-a-y.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8554 cdf35c40-ae34-48e0-9cc9-0c9da1808c22