block_self should not block all buttons when inside a members loop on one's own profile
This changeset reconfigures the block_self logic in BP_Button so that the
bp_is_my_profile() check is skipped when inside of a members loop. This allows
block_self logic to operate properly when inside of a members loop that happens
to appear on the logged-in user's profile page.

Also adds unit tests for the block_self param.

Fixes #5734

Props r-a-y for the initial patch.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8607 cdf35c40-ae34-48e0-9cc9-0c9da1808c22