xProfile: when updating a user profile, make sure the actions are passed the field's current (new) value.
Previously, we've been passing these actions the values that we've told xProfile to set, instead of the
field's actual (new) value. Most of the time, these will be identical, but with BP 2.0's xProfile field
type improvements, there are more lower-level places where the value could be filtered before the data's
written to the database.

In this situation, the field's actual new value could be different from what we've been returning in the
`xprofile_profile_field_data_updated` and `xprofile_updated_profile` actions.

See #5755

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8631 cdf35c40-ae34-48e0-9cc9-0c9da1808c22