Mentions: make `filter` behaviour match endpoint.
For pre-primed lookups, the regex was only matching matches at the
beginning of words, or after a space. BP’s Suggestions API has been
built to find matches of a value in any position.

The change removes the `^` character from the regex so that the
behaviour is the same. The highlighting still needs to be fixed.