Make sure autocomplete feature is available in Message compose screen
As the autocomplete javascripts are now registered hooking 'bp_enqueue_scripts' with a priority of 1 since r8559, the function {{{messages_add_autocomplete_js()}}} also needs to be hooked to 'bp_enqueue_scripts' but with a later priority (default one). Previously it was hooked to 'bp_actions' which is too early for the javascripts to be registered.

Fixes #5783

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8709 cdf35c40-ae34-48e0-9cc9-0c9da1808c22