Do not interfere with WordPress sign-up process when the user is created from the blog or the network administration.
On multisite config, make sure the activation email and the welcome email generated by WordPress are not modified by BuddyPress if the user was created from the User Administration panel. In this case, user needs to receive the built in activation link and/or the login and password generated by WordPress.

Props godavid33, r-a-y

Fixes #5789

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8922 cdf35c40-ae34-48e0-9cc9-0c9da1808c22