bp-legacy: Fix notice when replying to private group activity updates via AJAX.
When a reply is made to an activity update, we have to repopulate the
`$activities_template` global with the new activity object.

However, when we fetch the new private activity comment, we have to add a
check for the visibility of the item as well.  This currently isn't done,
which led to the notices and the non-existence of the activity object.

This commit addresses the issue by setting the `show_hidden` parameter to
`true` when fetching the new activity item, ensuring that the
`$activities_template->activity` object will be set.

Fixes #5862.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8989 cdf35c40-ae34-48e0-9cc9-0c9da1808c22