Core: only return `forums` from `bp_core_get_packaged_component_ids()` if using bbPress 1.
This otherwise causes the bbPress 2's subnav bar to not be rendered in templates due to the behaviour of `members/single/plugins.php`.

We would normally consider using `bp_is_active` for this kind of check, but because both bbPress 1 and (possibly, in the future) 2 could return true, we need a more reliable check for bbPress 1 only; see the ticket for full details.

See #5552. Fixes #5873, props r-a-y.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9012 cdf35c40-ae34-48e0-9cc9-0c9da1808c22