Core: Page title: remove component name from members component item pages. Additional screens introduced by other plugins have been similarily changes.
This simplifies the <title> for the member profile screens; the longer the title, the less likely it will fit in the narrow space afforded by modern web browsers. I expect most people will only have one tab open for a one member (for example); it's an edge case to have multiple tabs open on different parts of the same user's profile for most/general use.

See #5838, props sooskriszta for the initial patch.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9105 cdf35c40-ae34-48e0-9cc9-0c9da1808c22