Unit Tests, Groups: update the group factory to set the group's `creator_id` if one was not passed to the factory.
This is a quicker way of updating hundreds of unit tests where we've previously had a bug that was creating Groups without a valid group creator being set. See #5983 for related changes.

A few tests are counting the number of users in specific circumstances, so those eight tests have been individually updated so that they pass a valid `creator_id` without us having to refactor the entire test.

The `test_group_has_no_members` test has also been removed as it is impossible for a group to have no members. This test used to work because of this bug.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9120 cdf35c40-ae34-48e0-9cc9-0c9da1808c22