Build: revert r9150 ("have Travis-CI shallow clone the repo").
The idea was to reduce the amount of data Travis-CI had to download when it cloned the test prior to each test run.
However, there doesn't seem to be much evidence to suggest this helped speed up the running of the test suite in a meaningful way.

The problem with using depth 1 is that if we make multiple commits in quick succession before Travis-CI is able to
kick off a build for that commit, it will fail due to being unable to find the commit. This is because Travis-CI seems to check out a branch and then check out the exact commit. It's possible to get into a situation where the latest branch is newer than the commit Travis is trying to build against.

Thanks to https://github.com/Freeseer/freeseer/commit/84252a6afc1465265565dd72d17c0bc438ba0988 for figuring out the black magic behind these failures.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9199 cdf35c40-ae34-48e0-9cc9-0c9da1808c22