Introduce Member Types API.
BuddyPress member types are akin to WordPress's custom post types. Developers
can use `bp_register_member_type()` to register their types with BuddyPress,
and BP will automatically provide a number of pieces of functionality:

* Mechanisms for storing and fetching member types (as a WP taxonomy term), with full cache support.
* A 'member_type' argument for the `bp_has_members()`/`BP_User_Query` stack, which allows filtering member loops by member type.
* Admin UI for changing member types on Dashboard > Users > Community Profile (appears when member types have been registered).

We'll continue to build out more core member type functionality in future
versions of BuddyPress. In the meantime, this is a good starting point for BP
site implementations to have a shared infrastructure for storing and retrieving
this data.

See #6006.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9210 cdf35c40-ae34-48e0-9cc9-0c9da1808c22