Use toolbar for narrowest display.
We’re registering the BuddyPress nav to the WP Toolbar but we are
usually hiding it. On narrow displays, we unhide it, and hide the BP
nav bar for a simple usability solution.

Otherwise, we could end up with two nav bars, which is hard to
implement, and probably confusing.