Make sure paginate links are correctly set in case javascript is disabled
Adding extra query arguments to the 'base' parameter of the {{{paginate_links()}}} function when javascript is disabled is preventing the loops to be correctly paginated. In this particular case, the pagination is 'stuck' on the second page. To be sure the pagination is correctly set, we need to pass extra query arguments using the 'add_args' parameter of the {{{paginate_links()}}} function.

Fixes #5967

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9260 cdf35c40-ae34-48e0-9cc9-0c9da1808c22