bp-legacy: Add better detection for the group member search field.
Previously, if the activity component is disabled and a user is on a group
homepage, the AJAX search and select dropdown would not function properly.

This commit fixes this by setting the appropriate `object` and `scope`
parameters when the group member search field is detected.

Props vimes1984 for an initial patch.

Fixes #6033.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9305 cdf35c40-ae34-48e0-9cc9-0c9da1808c22