If a component's mapped Page's template is set in wp-admin, use that instead of BP's default template stack.
This allows people to override which page template BuddyPress injects its content into.

While this was previously possible by creating a specially-named template file in the theme, the technique was a little obscure. This alternate solution is user friendly and allows different container page templates on a per-component basis.

Fixes #6065

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9306 cdf35c40-ae34-48e0-9cc9-0c9da1808c22