Fix notice errors when an activity is generated for a post having no title.
Post types activities are using dynamic action strings when the activity is displayed in the stream and when the activity is saved into the database. In this particular case, once the post is published, we are "faking" an activity object to which we add two extra properties : the post_title and the post_url. To avoid switching blogs too many times, we are saving the post title into an activity meta. To know if we must build the action string for insert or display, we are checking the post_title property of the activity object. When a post is published without a title, the `empty()` check was not right. Changing for an `isset()` check prevents the notice errors.

props boonebgorges

Fixes #6119

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9369 cdf35c40-ae34-48e0-9cc9-0c9da1808c22