On multisite configs, make sure unspamming the user from the Users Network Administration screen is updating the .
When a user is marked as spam from the front end, then marked as ham from the network-users administration screen using the 'unspam' bulk action, we need to be sure the  field of the users table is updated so that BuddyPress effectively takes in account the user has been unspammed.

This commit also makes sure the user's cache has been cleared.

Fixes #5275

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9398 cdf35c40-ae34-48e0-9cc9-0c9da1808c22