BP-Legacy: add a new param to `locate_asset_in_stack` to allow the enqueued asset's `$handle`to be explicitly set.
Originally designed to load a template pack's main CSS and JS files in a theme compatibility manner, in r8686 `locate_asset_in_stack` was used to enqueue a JS file for password strength meters (registration and settings screens). This worked, but the `$handle`, dynamically constructed and passed to `wp_enqueue_script`, for the password strength JS file is likely to be the exact same handle as that used for a template pack's main JS file. This has the consequence of making the file(s) registered first unenqueue-able.

The new parameter introduced by this change allows us to explicitly set the name of an enqueued asset's handle, which avoids this problem. Technically the script handle for the password strength meter JS was never intentionally set, so we avoid most of the concerns of a backwards compatibility break, and now allow developers to actually unenqueue a template pack's CSS and JS.

Fixes #5972.
See also #5797.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9524 cdf35c40-ae34-48e0-9cc9-0c9da1808c22