Ensure that corresponding notifications are deleted when a message thread is deleted.
This means deleting the notifications for each message in the thread, for each
recipient on the thread. Previously, only a subset of these were being deleted.

This changeset removes a test that used a dummy thread_id, a testing
technique that doesn't work with the new looping technique used for
notification deletion. The new unit test is a proper replacement.

Props pareshradadiya for an initial patch.
Fixes #6329.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9744 cdf35c40-ae34-48e0-9cc9-0c9da1808c22