Improved plural forms for localizable strings.
Strings that are intended for use only in the case of a single item should not
be bundled with strings intended for use with plurals, as this pattern does not
work with languages that use the singular nominal form for numbers other than
one. See https://core.trac.wordpress.org/ticket/28502#comment:23.

This changeset also fixes a few places where a formatted string, rather than a
raw integer, was being passed as the `$count` value to `_n()`.

Props SergeyBiryukov.
Fixes #6368.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9765 cdf35c40-ae34-48e0-9cc9-0c9da1808c22