Convert second parameter of `bp_current_user_can()` to an `$args` array.
This makes the syntax simpler for the vast majority of use cases, where an
explicit `blog_id` doesn't need to be specified. It also makes it possible to
pass additional arguments, such as an `item_id`, to `bp_current_user_can()`.

An integer passed as the second parameter will trigger the backward-
compatibility layer.

Props thebrandonallen, r-a-y.
See #5121. Fixes #6501.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9957 cdf35c40-ae34-48e0-9cc9-0c9da1808c22