BuddyPress Uploader: avoid duplicate ids in warnings
The BuddyPress Uploader is displaying some warnings to inform the user when something went not great while trying to upload a file (eg: file is too big).
In case of multi uploads, we need to make sure these warnings are not using the same id, and to remove them all when a new upload starts.

Fixes #6590

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10049 cdf35c40-ae34-48e0-9cc9-0c9da1808c22