Detect if wp_mail has been re-defined, handle appropriately.
Most plugins that re-define wp_mail are integration with an external
mail service, often with their own templates, so we don’t want to
affect those out-of-the-box.

Likewise, HTML emails have to be filtered to be turned on, so if that’s
happened, we can be reasonably sure someone’s already customised the
WordPress email experience, and so we should get out of the way.