Groups Homes: make sure the action parameter of the canonical stack is reset to `home` in two specific cases.
When the Group does not use a custom `front.php` template:
(1) If the Activity component is not active, the home page displays the members of the Group.
(2) If the Activity component is active, the home page displays the activities of the Group.

In case (1), a user trying to directly access to `site.url/groups/single/members` must be redirected
to the home page.
In case (2), a user trying to directly access to `site.url/groups/single/activity` must be redirected to the home page.

Props jmarx75 (Congrats for your first ticket and patch, welcome on board!)

Fixes #6646

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10205 cdf35c40-ae34-48e0-9cc9-0c9da1808c22