bp-legacy: Only set activity cookies if they are not `null`.
When we upgraded jQuery.cookie to v1.4.1 in BuddyPress 2.1.0 (#5208),
`null` values now get set as cookies.  Beforehand, the older version of
jQuery.cookie() did not set `null` values as a cookie.

This caused an issue on the Activity directory page where the correct tab
was not being selected due to the invalid cookie value.

Fixes #6143.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10214 cdf35c40-ae34-48e0-9cc9-0c9da1808c22