bp-legacy: fix nonce and message when requesting private group membership
The AJAX handler for the “request private group membership” button was
sending the wrong nonce in its response when you leave a private group.
The button refreshes its label to read “request membership” but as the
nonce was wrong, only “-1” was printed to the screen. Nothing actually
happened.

By coincidence, this is one of those legacy parts of BuddyPress where
markup is shared between a template file and inside an AJAX callback
function in the PHP. While resolving that duplication is beyond the
scope of this immediate change, the AJAX version of the template has
been updated to match the template file (the only change is the button
label).

Fixes #5790

Props lakrisgubben

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10237 cdf35c40-ae34-48e0-9cc9-0c9da1808c22