xprofile admin: fix handling of slashes in xprofile group name/desc
Add filters to parts of the xprofile wp-admin templates so we're
running group description and name through the same output filters that
are used in the front-end template parts (for stripslashes).

Also adds a stripslashes to profile group name and description
before_save filters to fix the root issue. By keeping the existing
on-display stripslashes, we maintain backwards compatibility.

Fixes #6081

Props mahype, djpaul

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10239 cdf35c40-ae34-48e0-9cc9-0c9da1808c22