Tests: fix test failures on PHP7
PHP7 has changed its array sorting behaviour so that it is not
necessarily guaranteed to match previous versions of PHP. Brittle
tests, such as these, fail because of the change.

As a workaround, let's try using `assertEqualSets`, which is a
WordPress-provided testing assert that does a `sort` on each array
before comparing the arrays.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10260 cdf35c40-ae34-48e0-9cc9-0c9da1808c22