Blogs: Record activity item for blog post if it doesn't exist when a post comment is posted.
Fixes an issue where if "Allow activity stream commenting on blog and forum
posts" is enabled, the activity item for the post comment would not be
created because the activity item for the blog post does not exist.

This might occur if the "Site Tracking" component is enabled after posts
have already been published.

Commit also includes a unit test.

Fixes #6530.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10277 cdf35c40-ae34-48e0-9cc9-0c9da1808c22