Emails: introducing `bp_send_email`, which is our replacement for `wp_mail`.
As `wp_mail` is a pluggable function, if we detect it has been re-declared, or if something has filtered to enable HTML emails, BuddyPress will fallback and let `wp_mail` send the email for us. This maintains compatibility for customised sites until mail delivery plugins are updated to work with BuddyPress (or new plugins are written).

See #6592. Props timersys, mercime, boonebgorges, hnla, DJPaul.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10478 cdf35c40-ae34-48e0-9cc9-0c9da1808c22