Emails: use local instance of WordPress' PHPMailer class.
Relying on the global is okay, but initialising it (if it is unset) touches a WordPress global which doesn't have a direct API. While this is probably safe today, it may not be reliable in the future. After consideration, especially considering BuddyPress' efforts not to introduce or use globals directly, now we instead create a static instance of the `PHPMailer` object inside the `BP_Email` class.

The original approach came from WordPress' `wp_mail()` which, I assume, once found a situation where the global wasn't set.

For #6592

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10490 cdf35c40-ae34-48e0-9cc9-0c9da1808c22