Sort activity actions when they're added, instead of when they're fetched.
Sorting activity actions can be an expensive process, so should not be done
any more frequently than necessary. And it's only necessary to sort when a
new action has been added. So instead of sorting in `bp_activity_get_actions()`,
we now do it in `bp_activity_set_action()`.

Props boonebgorges, imath.
Fixes #6865.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10515 cdf35c40-ae34-48e0-9cc9-0c9da1808c22