Tools, Emails: empty email type taxonomy when running the email reinstall tool.
If a 3rd party plugin adds email types which aren't cleared up when that plugin is deactivated, we could end up with orphaned terms which would continue to show up in the email "sitation" metabox. As we hide the email type taxonomy management screen in wp-admin by default, this makes it very hard for the average user to remove those unused distracting terms.

Props modemlooper

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10583 cdf35c40-ae34-48e0-9cc9-0c9da1808c22