Profile: Improve validation in is_valid().
When checking a submitted array of values,
ensure that the `foreach` breaks when an
invalid value is encountered. Previously,
the `foreach` returned the validity of the
last element in the array.

Also update unit tests so that invalid values
appear in various positions in the array of
values to check.

Fixes #6884.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10592 cdf35c40-ae34-48e0-9cc9-0c9da1808c22