Make sure the BuddyPress Email template can be loaded within the WordPress Administration.
`bp_locate_template()` is only loading templates if `WP_USE_THEMES` is defined or if an Ajax request is done, which is not the case inside a regular Administration page. As the email template is a required parameter for the BP_Email class, this was preventing any email to be sent from the WordPress Administration (eg: the activation email using the "resend" action from the Signups screen).
To fix this we are introducing a new filter the BP_Email class is using to get the email template.

Fixes #6912

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10594 cdf35c40-ae34-48e0-9cc9-0c9da1808c22