Multisite: make sure regular users can create new sites when the blogs component is not active.
`strpos()` function was used the wrong way around, as a result clicking on the "Add new" button at the top of the "My Sites" dashboard page was redirecting the user to root blog home. Using the `strpos()` function the right way is making sure the user can access to the WordPress signup page and create a new site.

Fixes #6178

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10665 cdf35c40-ae34-48e0-9cc9-0c9da1808c22