Introduce screen notifications for activity comments and activity comment replies
Users will be notified about comments made on their activities and about replies made on their activity comments.

If there is only one notification of the above types, the WP Admin Bar bubble will link to the single activity. Clicking on this link will mark the notification as read.
If there are more than one, we use the notifications screen and an extra parameter to filter the notifications according to the type so that only these notifications are displayed. Clicking on the listed links will mark the notification as read.

Props dcavins, boonebgorges, DJPaul

Fixes #6057

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10715 cdf35c40-ae34-48e0-9cc9-0c9da1808c22