Activity: Fix issue with activity permalink access when the legacy activity query is enabled.
In #7113, we fixed a bug where anybody's activity item could be displayed
under a member profile.  The fix (r10880) does a strict type check on
`$activity->user_id`.

We expect `$activity->user_id` to be an integer, however it can be a
numeric string if the legacy activity query is enabled with:
`add_filter( 'bp_use_legacy_activity_query', '__return_true' );`

Since the check will always fail in this instance, members will not be able
to access activity permalinks.

We unfortunately missed this specific instance in r10853.

This commit addresses this problem.

See #6977. Fixes #7146 (trunk).

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10908 cdf35c40-ae34-48e0-9cc9-0c9da1808c22