XProfile: Return array if we pass `$user_ids` as an array in `BP_XProfile_ProfileData::get_value_byid().
Previously, in r10915, we changed the behavior to calculate whether we
should return a single result by counting the number of user IDs.  This
created a PHP notice for some that are relying on the return value to be
an array instead of a string.

This commit reverts this behavior and renames the `$is_single` variable
to `$return_single_result` to better explain what is happening.

See r10915 (trunk).

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10921 cdf35c40-ae34-48e0-9cc9-0c9da1808c22