Core: stash the unsubscribe email salt as a static var in `bp_get_default_options()`.
In r10941, an email unsubscribe feature was introduced. This added a new site option that stores the verification salt. For new installations, we added the variable to `bp_add_options`, and for upgrades, in a new function. We generate the value by using a randomly-generated string.

On WordPress older than 4.4, `wp_rand`'s implementation was causing infinite loops because it adds its own site option, and BuddyPress has some cache clearing logic in `bp_core_clear_root_options_cache` which is called whenever options are added or updated. WordPress 4.4 re-implemented `wp_rand` and doesn't cause a loop.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10945 cdf35c40-ae34-48e0-9cc9-0c9da1808c22