Only delete the "new_blog" activity when its author is removed from the Blog.
Before this commit, as soon as any contributor was removed from the blog, the "new_blog" activity was also removed although the removed user was not the author of the activity.
It is no more the case. This activity will only be deleted if the user who created the blog is removed from it.
This commit also make sure this activity is not deleted twice as some duplicate code has been left when moving the activity code inside the `bp-blogs-activity.php` file during the 2.6 dev-cycle.

Fixes #7199 (trunk)

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10971 cdf35c40-ae34-48e0-9cc9-0c9da1808c22