Build: `scss-lint` requires Ruby >= 2.2.0
Travis-CI has started to fail because an underlying `scss-lint` library
we require, now requires a newer version of Ruby (at time of writing,
Ruby 1.9.3 is pre-installed on Travis-CI PHP boxes).

This change updated the Travis-CI build script to use the newer version
of Ruby. It also adds the Ruby gems cache folder to the Travis-CI job
cache to try to minimise the longer set-up time that this change has
incurred.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11040 cdf35c40-ae34-48e0-9cc9-0c9da1808c22