Core: Fix fatal error with `bp_core_get_iso8601_date()` when an invalid date string is passed.
This error occurred on the Members Directory and selecting the
"Alphabetical" filter when attempting to parse an inactive user's last
activity.  Inactive users passes the "Never active" string to
`bp_core_get_iso8601_date()` and the `DateTime` class causes a fatal
error on invalid date strings, unless we catch the exception, which this
commit is now doing.

See #5757.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11097 cdf35c40-ae34-48e0-9cc9-0c9da1808c22