Add multisite-aware wrapper for get_objects_in_term().
Add wrapper functions for working with BuddyPress taxonomies via
`get_objects_in_term()`. When working with BuddyPress objects like
users, we need to be sure that we’re referring to the taxonomies and
terms for the site on the network where the relevant taxonomy is
registered.

Props dcavins, boonebgorges, offereins.

See #7263.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11118 cdf35c40-ae34-48e0-9cc9-0c9da1808c22