Friends tests: Set date_created to ensure reliable ordering.
Some of the tests for friendship functionality use `assertEquals()`,
which requires that the elements of arrays be in the same order and
each value be equal. In some cases, the order of the elements returned
by functions like `friends_get_friendship_request_user_ids()` was
unreliable, probably because phpunit was working quickly enough that
the `date_created` was indeterminate. This change set manually sets the
earlier friendship back in time so that the return values can be
checked reliably.

See #6978.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11125 cdf35c40-ae34-48e0-9cc9-0c9da1808c22