Blogs: Use WordPress' Site Icons as blog avatars.
If the current install is using WordPress 4.3+, we'll now use WordPress'
Site Icon as the site avatar in the Sites Directory:
https://codex.wordpress.org/Creating_a_Favicon#WordPress_Version_4.3_or_later

If no site icon is available, we'll fallback to the site owner's user
avatar as before.  Under the hood, we are mirroring the site icon URLs to
blogmeta whenever a site icon is updated or deleted.

Devs can disable this functionality with the following filter:
`add_filter( 'bp_is_blogs_site-icon_active', '__return_false' );`

Fixes #6544.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11150 cdf35c40-ae34-48e0-9cc9-0c9da1808c22