Tests: rely on WP Core's version of `go_to()`.
Our version of `go_to()` is very messy, and since we copied it a few years ago, it has been refactored in WordPress.
We can remove the duplication which makes for a cleaner method, and rely on the upstream version.

This also helps avoid false positives in tools that check for use of deprecated WordPress functions.
There was a conditional block in the old `go_to()` that was for compatibility with WordPress <3.9.

Fixes #7326

Props netweb, DJPaul

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11254 cdf35c40-ae34-48e0-9cc9-0c9da1808c22