Groups: Invitations settig unsync between front end and back end.
Invitation status may be unsynced for older groups, because there is no fallback used in group editing wp-admin area. We call groupmeta directly there for a value (which can be empty), but should use `bp_group_get_invite_status()`` instead, so it's filterable and is the same logic as on front-end.

Fixes #7399.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11305 cdf35c40-ae34-48e0-9cc9-0c9da1808c22