Discard activity filter values that are invalid for the current context.
Activity filters are stored in a cookie that persists across
activity pages. But some filter values are not valid in all contexts,
such as `new_member` (valid in the sitewide feed, but not in a group).
This changeset ensures that filter values sent via AJAX are discarded
if they aren't valid in the current context.

Props lakrisgubben.
Fixes #4062.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11356 cdf35c40-ae34-48e0-9cc9-0c9da1808c22