CI: update Travis-CI configuration.
* Remove WordPress 4.4 from matrix (we now test only WordPress 4.5+).
* Remove PHP 5.4, 5.5, 7.0, 7.2 (RC), and nightly, from matrix (we now test only PHP 5.3, 5.6, 7.1).
* Relocate `node_modules` folder and cache Node for quicker overall runtime.

The plan is that we test against our supported minimum PHP version, and the latest, as well as the last minor version of an old major branch.
Testing against non-release versions of PHP gives us very little benefit (any issues we are likely to find will cause problems in WordPress, and that team are more competent than us to report such issues upstream to PHP).


git-svn-id: https://buddypress.svn.wordpress.org/trunk@11682 cdf35c40-ae34-48e0-9cc9-0c9da1808c22