Nouveau: Add ruleset to support signup BP checks for in use names/emails
Commit adds rule to ensure invalid submissions where username or email already exists are styled with the visual warning indicator.
This ensures that a technically valid input value styled on the CSS pseudo attribute as valid may be overwritten for a BP error response.

This supports recent changes to nouveau_signup_form to add classes on error return response in r11770

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11775 cdf35c40-ae34-48e0-9cc9-0c9da1808c22