Navigation: Restricted subnav menu pages should redirect instead of throwing a 404.
Back in BuddyPress 2.6.0, we refactored the way we register menu items with
the BP Navigation API (see #6534).

The problem is we missed an instance during the refactor.  This caused
restricted subnav pages (those set with the `'user_has_access'` parameter
to false in `bp_core_new_subnav_item()`) to 404 instead of being redirected
away.

Commit fixes this problem.

See #7659.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11814 cdf35c40-ae34-48e0-9cc9-0c9da1808c22