Pass additional $args to 'bp_get_new_group_invite_friend_list()`
This function renders markup for the looped item defaulting to 'li' but leaves it to the template to provide the parent element.

Commit provides two new args 'before' & 'after' to allow the function to return the complete element markup if desired, simplifies templates having to render markup and checks for empty function return.

Props DJPaul.

Fixes #7656

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11828 cdf35c40-ae34-48e0-9cc9-0c9da1808c22